move components to the right place
