+++ src/modules/notifications/components/in/In.vue
@@ -24,11 +24,7 @@
         <div class="truncate pr-sm-2">
           <div class="d-flex align-center">
             <div class="mr-1">From:</div>
-            <EllipsisBlock
-              max-width="200px"
-              class="monospace"
-              :text="address"
-            />
+            <mew-transform-hash :hash="address" />
           </div>
 
           <div>Abount: {{ amount }}</div>
@@ -49,7 +45,7 @@
             class="d-block d-sm-flex align-start justify-space-between"
           >
             <div class="pr-2 font-weight-bold">{{ d.label }}</div>
-            <EllipsisBlock
+            <mew-transform-hash
               v-if="d.ellipsis"
               max-width="200px"
               class="monospace text-sm-right"
@@ -67,12 +63,7 @@
 </template>
 
 <script>
-import EllipsisBlock from '@/components/ellipsisBlock/EllipsisBlock';
-
 export default {
-  components: {
-    EllipsisBlock
-  },
   props: {
     address: {
       type: String,
